// src/api/model/monoPaymentProcessRequest.ts
/**
 * Generated by orval v7.7.0 🍺
 * Do not edit manually.
 * FinMan API
 * OpenAPI spec version: 1.0.0
 */
import type { MonoPaymentProcessRequestFromDate } from "./monoPaymentProcessRequestFromDate";
import type { MonoPaymentProcessRequestToDate } from "./monoPaymentProcessRequestToDate";
import type { MonoPaymentProcessRequestMode } from "./monoPaymentProcessRequestMode";

export interface MonoPaymentProcessRequest {
  /** ID користувача Mono */
  mono_user_id: number;
  /** Початкова дата (формат: YYYY-MM-DD) */
  from_date?: MonoPaymentProcessRequestFromDate;
  /** Кінцева дата (формат: YYYY-MM-DD) */
  to_date?: MonoPaymentProcessRequestToDate;
  /** Режим обробки платежів */
  mode?: MonoPaymentProcessRequestMode;
  /** ID користувача */
  user_id: number;
}